order service refactor

+++ src/main/java/com/brokerage/event/CreateOrderEvent.java
@@ -1,5 +1,6 @@
 package com.brokerage.event;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record CreateOrderEvent(UUID orderId, UUID customerId, String assetName, String orderSide, int size, double price) {}
+public record CreateOrderEvent(UUID orderId, UUID customerId, String assetName, String orderSide, int size, BigDecimal price) {}
